Fix navbar dropdown styling
Disabled buttons were having a weird behaviour when changing the cursor on
hover. Now they uniformly make the cursor "not-allowed". This has the
aditional upside that the "disabled" attribute is no longer required on
MenuItems.

This commit also changes slightly the text color in the dropdown, to make it
comply with the WCAG AA contrast recommendations when hovered.