Enforce "carousel" for consistency
Both "carrousel" and "carousel" were used for naming the left sidebar in
the application. All matches of "carrousel" have been changed to
"carousel" for the sake of consistency.

"carousel" has been chosen over "carrousel" because it's the original
French word and because it seems to be the most used variant.